requirements.txt
azure-storage-blob==12.19.0
pandas==2.3.1
requests==2.31.0
python-dotenv==1.0.0
jupyter==1.0.0
matplotlib==3.8.0
